Support customization of numOfProbe and probeInterval when externaltrafficpolicy is local
