+++ tests/e2e/network/service_annotations.go
@@ -838,24 +838,82 @@ var _ = Describe("Service with annotation", Label(utils.TestSuiteLabelServiceAnn
 		Expect(err).NotTo(HaveOccurred())
 
 		By("Validating health probe configs")
-		var numberOfProbes *int32
-		var intervalInSeconds *int32
 		for _, probe := range targetProbes {
 			if probe.ProbeThreshold != nil {
-				numberOfProbes = probe.ProbeThreshold
+				utils.Logf("Validating health probe config numberOfProbes")
+				Expect(*probe.ProbeThreshold).To(Equal(int32(3)))
 			}
 			if probe.IntervalInSeconds != nil {
-				intervalInSeconds = probe.IntervalInSeconds
+				utils.Logf("Validating health probe config intervalInSeconds")
+				Expect(*probe.IntervalInSeconds).To(Equal(int32(10)))
 			}
+			Expect(probe.Protocol).To(Equal(network.ProbeProtocolHTTP))
 		}
-		utils.Logf("Validating health probe config numberOfProbes")
-		Expect(*numberOfProbes).To(Equal(int32(3)))
-		utils.Logf("Validating health probe config intervalInSeconds")
-		Expect(*intervalInSeconds).To(Equal(int32(10)))
-		utils.Logf("Validating health probe config protocol")
-		Expect(len(targetProbes)).To(Equal(expectedTargetProbesCount))
-		for _, targetProbe := range targetProbes {
-			Expect(targetProbe.Protocol).To(Equal(network.ProbeProtocolHTTP))
+
+		By("Changing ExternalTrafficPolicy of the service to Local")
+		expectedTargetProbesCount = 1
+		var service *v1.Service
+		utils.Logf("Updating service " + serviceName + " in namespace " + ns.Name)
+		retryErr := retry.RetryOnConflict(retry.DefaultRetry, func() error {
+			service, err = cs.CoreV1().Services(ns.Name).Get(context.TODO(), serviceName, metav1.GetOptions{})
+			if err != nil {
+				return err
+			}
+			service.Spec.ExternalTrafficPolicy = v1.ServiceExternalTrafficPolicyTypeLocal
+			_, err = cs.CoreV1().Services(ns.Name).Update(context.TODO(), service, metav1.UpdateOptions{})
+			return err
+		})
+		Expect(retryErr).NotTo(HaveOccurred())
+		utils.Logf("Successfully updated LoadBalancer service " + serviceName + " in namespace " + ns.Name)
+
+		By("Getting updated service object from server")
+		retryErr = retry.RetryOnConflict(retry.DefaultRetry, func() error {
+			service, err = cs.CoreV1().Services(ns.Name).Get(context.TODO(), serviceName, metav1.GetOptions{})
+			if err != nil {
+				return err
+			}
+			if service.Spec.HealthCheckNodePort == 0 {
+				return fmt.Errorf("service HealthCheckNodePort is not updated")
+			}
+			return nil
+		})
+		Expect(retryErr).NotTo(HaveOccurred())
+
+		err = wait.PollImmediate(5*time.Second, 60*time.Second, func() (bool, error) {
+			lb = getAzureLoadBalancerFromPIP(tc, publicIPs[0], tc.GetResourceGroup(), "")
+			targetProbes = []*network.Probe{}
+			for i := range *lb.LoadBalancerPropertiesFormat.Probes {
+				probe := (*lb.LoadBalancerPropertiesFormat.Probes)[i]
+				utils.Logf("One probe of LB is %q", *probe.Name)
+				probeSplit := strings.Split(*probe.Name, "-")
+				Expect(len(probeSplit)).NotTo(Equal(0))
+				probeSplitID := probeSplit[0]
+				if len(probeSplit) > 1 &&
+					(probeSplit[len(probeSplit)-1] == "IPv4" || probeSplit[len(probeSplit)-1] == "IPv6") {
+					probeSplitID += "-" + probeSplit[len(probeSplit)-1]
+				}
+				for _, id := range ids {
+					if id == probeSplitID {
+						targetProbes = append(targetProbes, &probe)
+					}
+				}
+			}
+
+			return len(targetProbes) == expectedTargetProbesCount, nil
+		})
+		Expect(err).NotTo(HaveOccurred())
+
+		By("Validating health probe configs")
+		for _, probe := range targetProbes {
+			if probe.ProbeThreshold != nil {
+				utils.Logf("Validating health probe config numberOfProbes")
+				Expect(*probe.ProbeThreshold).To(Equal(int32(5)))
+			}
+			if probe.IntervalInSeconds != nil {
+				utils.Logf("Validating health probe config intervalInSeconds")
+				Expect(*probe.IntervalInSeconds).To(Equal(int32(15)))
+			}
+			Expect(probe.Protocol).To(Equal(network.ProbeProtocolHTTP))
 		}
 	})
 
@@ -1438,9 +1496,13 @@ func getAzureLoadBalancerFromPIP(tc *utils.AzureTestClient, pip, pipResourceGrou
 	return &lb
 }
 
-func createAndExposeDefaultServiceWithAnnotation(cs clientset.Interface, ipFamily utils.IPFamily, serviceName, nsName string, labels, annotation map[string]string, ports []v1.ServicePort) []string {
+func createAndExposeDefaultServiceWithAnnotation(cs clientset.Interface, ipFamily utils.IPFamily, serviceName, nsName string, labels, annotation map[string]string, ports []v1.ServicePort, customizeFuncs ...func(*v1.Service) error) []string {
 	utils.Logf("Creating service " + serviceName + " in namespace " + nsName)
 	service := utils.CreateLoadBalancerServiceManifest(serviceName, annotation, labels, nsName, ports)
+	for _, customizeFunc := range customizeFuncs {
+		err := customizeFunc(service)
+		Expect(err).NotTo(HaveOccurred())
+	}
 	_, err := cs.CoreV1().Services(nsName).Create(context.TODO(), service, metav1.CreateOptions{})
 	Expect(err).NotTo(HaveOccurred())
 	utils.Logf("Successfully created LoadBalancer service " + serviceName + " in namespace " + nsName)
